requirements.txt
Django==2.1.5
django-cors-headers==2.4.0
djangorestframework==3.9.1
psycopg2-binary==2.7.6.1
pytz==2018.9
